Move LaTeX things into a new file
